requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.7
fastapi==0.104.0
filelock==3.12.4
frozenlist==1.4.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
lit==17.0.3
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.2
numba==0.58.1
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230918
pydantic==2.4.2
pydantic_core==2.10.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
websockets==12.0
yarl==1.9.2
